Query builder: Struct fields to override embedded fields
When an outer struct defines an SQL column name that matches one that is also defined in an embedded struct, only the field in the outermost struct will be included in a query. This resolves an issue where overridden columns would result in multiple fields of the same name being set (like in an UPDATE statement), resulting in a Postgres query error.